add pfp context to update all instances of usePfp hook on update
